Fully functional Upgrade feature.
Created 5 upgrades and an upgrade page. The page has 5 buttons to purchase upgrades, text descriptions for each upgrade, and a dynamically changing text field for Cost/PerSecond/Amount of upgrade.

Added an automation flag to "pause" updating the meal count when on Upgrades page.

Created parallel arrays that store base values for the Upgrades.java.

Created automation for upgrades to update meals using a thread.

Added a lot of get/update/check methods in MealLogic.autoIncreaseMeals.

Added a "totalMeals" in mealPersistence.

Fixed the Statistics activity to get the correct totalMeal amount and display it.

Added Costs to each instance of Upgrade domain objects.