Created actual persistent storage for the Meals however it is extremely laggy and needs to be refactored. There are very very very frequent database accesses.
Added the dependencies in the build.gradle file for hsqldb.

Moved the onCreate method to the top in MainActivity because I could never find it. Added a mealLogic.initializeDB() method that should/might be temporary

Added the PersistenceException class from the sample project to throw around exceptions.

SC.script now initializes the Meals table to default values of 0 for curr/total meals and numclicks.